+++ tekore/_model/currently_playing.py
@@ -1,4 +1,4 @@
-from typing import Optional, Union
+from typing import Optional, Union, List
 from dataclasses import dataclass
 
 from .context import Context
@@ -6,7 +6,7 @@ from .device import Device
 from .track import FullTrack
 from .local import LocalTrack
 from .episode import FullEpisode
-from .serialise import Model, StrEnum
+from .serialise import Model, StrEnum, ModelList
 
 
 class CurrentlyPlayingType(StrEnum):
@@ -58,6 +58,13 @@ item_type = {
 }
 
 
+def _parse_playback_item(item: dict):
+    if item.get('is_local', False) is True:
+        return LocalTrack(**item)
+    else:
+        return item_type[item['type']](**item)
+
+
 @dataclass(repr=False)
 class CurrentlyPlaying(Model):
     """
@@ -84,10 +91,7 @@ class CurrentlyPlaying(Model):
         if self.context is not None:
             self.context = Context(**self.context)
         if self.item is not None:
-            if self.item.get('is_local', False) is True:
-                self.item = LocalTrack(**self.item)
-            else:
-                self.item = item_type[self.item['type']](**self.item)
+            self.item = _parse_playback_item(self.item)
 
 
 @dataclass(repr=False)
@@ -102,3 +106,15 @@ class CurrentlyPlayingContext(CurrentlyPlaying):
         super().__post_init__()
         self.device = Device(**self.device)
         self.repeat_state = RepeatState[self.repeat_state]
+
+
+@dataclass(repr=False)
+class Queue(Model):
+    """Playback queue."""
+
+    currently_playing: Union[FullTrack, LocalTrack, FullEpisode, None]
+    queue: List[FullTrack]
+
+    def __post_init__(self):
+        self.currently_playing = _parse_playback_item(self.currently_playing)
+        self.queue = ModelList(_parse_playback_item(i) for i in self.queue)